:bricks: Use the same proxy for the different protocols
Have one reverse proxy per node, but use URL prefix to route to the
different port/protocols of the node.

Proxy addresses:
- https://canto-validator-nginx-reverse-proxy-node1-testnet-3mid33wd4a-uc.a.run.app/
- https://canto-validator-nginx-reverse-proxy-node2-testnet-3mid33wd4a-uc.a.run.app/

Currently available root routes:
- /api/ -> node:1317
- /rpc/ -> node:26656
- /evm_rpc/ -> node:8445

Update the `trust_height` so the nodes resync as they got destroyed.
Ideally this should be automated, probably as part of the `entrypoint.sh`.