:sparkles: Sentry nodes
Setup two sentry nodes following this guide:
https://forum.cosmos.network/t/sentry-node-architecture-overview/454

The validator node ports are now all firewalled/closed on the external IP
except for SSH. The sentry nodes will communicate with the validators
using the validator internal/private VPC IP.
We're now importing the `node_key.json` on both validators and sentries so
that the node ID is known upfront.
This is needed the validator `persistent_peers` setting and the sentry
`private_peer_ids`.

Note the validators still have a public IP at the moment for convenience.